+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## v0.2.0 (2025-11-26)
+
+- `renderTrixContentWithEmbeds` is now `renderTrixContent`
+- Adds the following plugins which can be disabled through the `options` param of `renderTrixContent`:
+  - `codeBlock`
+  - `images`
+  - `inlineCode`
+- Adds `registerTrixPlugin` function allowing registration of 3rd party plugins
+
 ## v0.1.0 (2025-11-14)
 
 - Adds YouTube embed rendering